[Enhancement]avoid Aws::Client::ClientConfiguration default ctor which has a remote access
